Ignore the pull request from my separate branch
Reviewing Nicholas' code and adding basic functionality to the elevator subsystem. This included adding to the periodic function and creating a constructor that sets the motors' default brake mode to idle. I did some small things as well, such as setting the channels of each limit switch to a different number. I also added more comments to make the code more readable. Next steps would be testing with the actual robot and elevator.